+++ src/main/kotlin/org/jetbrains/research/testspark/actions/TestSparkAction.kt
@@ -12,6 +12,7 @@ import org.jetbrains.research.testspark.actions.template.PanelFactory
 import org.jetbrains.research.testspark.bundles.TestSparkLabelsBundle
 import org.jetbrains.research.testspark.display.TestSparkIcons
 import org.jetbrains.research.testspark.helpers.getCurrentListOfCodeTypes
+import org.jetbrains.research.testspark.services.LLMTestSampleService
 import org.jetbrains.research.testspark.services.SettingsApplicationService
 import org.jetbrains.research.testspark.tools.Manager
 import org.jetbrains.research.testspark.tools.evosuite.EvoSuite
@@ -29,7 +30,6 @@ import javax.swing.JFrame
 import javax.swing.JLabel
 import javax.swing.JPanel
 import javax.swing.JRadioButton
-import org.jetbrains.research.testspark.services.LLMTestSampleService
 
 /**
  * Represents an action to be performed in the TestSpark plugin.
